bug fix before update android studio

+++ app/src/main/java/com/dev/nicola/allweather/model/ForecastDay.java
@@ -7,10 +7,10 @@ public class ForecastDay {
 
     private String date;
     private String condition;
-    private double temperature;
+    private int temperature;
     private int icon;
 
-    public ForecastDay(String date, String condition, double temperature, int icon) {
+    public ForecastDay(String date, String condition, int temperature, int icon) {
         this.date = date;
         this.condition = condition;
         this.temperature = temperature;
@@ -34,7 +34,7 @@ public class ForecastDay {
         this.condition = condition;
     }
 
-    public double getTemperature() {
+    public int getTemperature() {
         return temperature;
     }
